refactor(ingestion): split framework_edges.py into per-framework handlers
Break the 1311-line framework_edges.py monolith into an
ingestion/framework_edges/ package — one module per framework
(django/fastapi/flask/aspnet/rails/laravel/spring/express/go/rust/typo3)
plus pytest_edges (conftest) and base.py for the shared primitives.

- base.py holds read_text, _add_edge_if_new, the class/function name->file
  maps, and a FrameworkHandler protocol (detect()/add_edges()).
- __init__.py is the facade: it re-exports add_framework_edges (the only
  externally-imported name) and now dispatches by iterating an ordered
  _HANDLERS list. Order is preserved so the first-wins _add_edge_if_new
  semantics are byte-for-byte identical.
- Collapse the _read_text / _read_cs_text variants and the rails/laravel
  inline route reads into base.read_text(parsed, encoding="utf-8");
  aspnet passes utf-8-sig for BOM files. The fastapi/flask inline reads
  use a different encoding + exception surface and are left inline.

Behaviour-preserving mechanical move. Ruff + mypy clean on the new
package; tests/unit/ingestion/ (632) and full tests/unit/ (2388) green.
